+++ pkg/services/mysqldb/types_dbms_only.go
@@ -1,0 +1,38 @@
+package mysqldb
+
+import "github.com/Azure/open-service-broker-azure/pkg/service"
+
+type dbmsOnlyMysqlInstanceDetails struct {
+	ARMDeploymentName          string `json:"armDeployment"`
+	ServerName                 string `json:"server"`
+	AdministratorLoginPassword string `json:"administratorLoginPassword"`
+	FullyQualifiedDomainName   string `json:"fullyQualifiedDomainName"`
+	EnforceSSL                 bool   `json:"enforceSSL"`
+}
+
+func (
+	d *dbmsOnlyManager,
+) GetEmptyProvisioningParameters() service.ProvisioningParameters {
+	return &ServerProvisioningParameters{}
+}
+
+func (
+	d *dbmsOnlyManager,
+) GetEmptyUpdatingParameters() service.UpdatingParameters {
+	return &UpdatingParameters{}
+}
+func (
+	d *dbmsOnlyManager,
+) GetEmptyInstanceDetails() service.InstanceDetails {
+	return &dbmsOnlyMysqlInstanceDetails{}
+}
+func (
+	d *dbmsOnlyManager,
+) GetEmptyBindingParameters() service.BindingParameters {
+	return &BindingParameters{}
+}
+func (
+	d *dbmsOnlyManager,
+) GetEmptyBindingDetails() service.BindingDetails {
+	return &mysqlBindingDetails{}
+}
